+++ Changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+###
+- Added AgendaItems to the Agenda object.
+
 ## [v0.2.2] - 2021-01-19
 
 ### Added